# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CustomReport {
  id: string;
  cafeId: string;
  ownerId?: string;
  name: string;
  description?: string | null;
  dateRangeMonths: number;
  includeInIm: boolean;
  incomeCount?: number;
  expenseCount?: number;
  createdAt?: string;
  updatedAt?: string;
}

export interface CreateCustomReportBody {
  cafeId: string;
  name: string;
  description?: string | null;
  dateRangeMonths?: number;
}

export interface UpdateCustomReportBody {
  name?: string;
  description?: string | null;
  dateRangeMonths?: number;
  includeInIm?: boolean;
}

export interface CustomReportLineItem {
  id: string;
  reportId: string;
  kind: string;
  label: string;
  source: string;
  xeroAccountId?: string | null;
  xeroAccountName?: string | null;
  sortOrder?: number;
  createdAt?: string;
}

export interface CustomReportLineItemInput {
  kind: string;
  label: string;
  source: string;
  xeroAccountId?: string | null;
  xeroAccountName?: string | null;
  sortOrder?: number;
}

export interface CustomReportMonthRow {
  month: string;
  income: number;
  expenses: number;
  net: number;
}

export type CustomReportDataTotals = {
  income?: number;
  expenses?: number;
  net?: number;
};

export type CustomReportDataGrowth = {
  momPct?: number | null;
  popPct?: number | null;
};

export type CustomReportDataPeriod = {
  fromDate?: string;
  toDate?: string;
  months?: number;
};

export interface CustomReportData {
  months: CustomReportMonthRow[];
  totals: CustomReportDataTotals;
  growth: CustomReportDataGrowth;
  period: CustomReportDataPeriod;
}

export type ListCustomReportsParams = {
cafeId: string;
};

export type ListCustomReports200 = {
  reports?: CustomReport[];
};

export type CreateCustomReport201 = {
  report?: CustomReport;
};

export type UpdateCustomReport200 = {
  report?: CustomReport;
};

export type DeleteCustomReport200 = {
  ok?: boolean;
};

export type GetCustomReportLineItems200 = {
  items?: CustomReportLineItem[];
};

export type ReplaceCustomReportLineItemsBody = {
  items?: CustomReportLineItemInput[];
};

export type ReplaceCustomReportLineItems200 = {
  items?: CustomReportLineItem[];
};