# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary List custom financial reports for a cafe
 */
export const ListCustomReportsQueryParams = zod.object({
  "cafeId": zod.coerce.string()
})

export const ListCustomReportsResponse = zod.object({
  "reports": zod.array(zod.object({
  "id": zod.string(),
  "cafeId": zod.string(),
  "ownerId": zod.string().optional(),
  "name": zod.string(),
  "description": zod.string().nullish(),
  "dateRangeMonths": zod.number(),
  "includeInIm": zod.boolean(),
  "incomeCount": zod.number().optional(),
  "expenseCount": zod.number().optional(),
  "createdAt": zod.string().optional(),
  "updatedAt": zod.string().optional()
})).optional()
})


/**
 * @summary Create a new custom financial report
 */
export const CreateCustomReportBody = zod.object({
  "cafeId": zod.string(),
  "name": zod.string(),
  "description": zod.string().nullish(),
  "dateRangeMonths": zod.number().optional()
})


/**
 * @summary Update report metadata
 */
export const UpdateCustomReportParams = zod.object({
  "id": zod.coerce.string()
})

export const UpdateCustomReportBody = zod.object({
  "name": zod.string().optional(),
  "description": zod.string().nullish(),
  "dateRangeMonths": zod.number().optional(),
  "includeInIm": zod.boolean().optional()
})

export const UpdateCustomReportResponse = zod.object({
  "report": zod.object({
  "id": zod.string(),
  "cafeId": zod.string(),
  "ownerId": zod.string().optional(),
  "name": zod.string(),
  "description": zod.string().nullish(),
  "dateRangeMonths": zod.number(),
  "includeInIm": zod.boolean(),
  "incomeCount": zod.number().optional(),
  "expenseCount": zod.number().optional(),
  "createdAt": zod.string().optional(),
  "updatedAt": zod.string().optional()
}).optional()
})


/**
 * @summary Delete a custom report
 */
export const DeleteCustomReportParams = zod.object({
  "id": zod.coerce.string()
})

export const DeleteCustomReportResponse = zod.object({
  "ok": zod.boolean().optional()
})


/**
 * @summary Get line items for a report
 */
export const GetCustomReportLineItemsParams = zod.object({
  "id": zod.coerce.string()
})

export const GetCustomReportLineItemsResponse = zod.object({
  "items": zod.array(zod.object({
  "id": zod.string(),
  "reportId": zod.string(),
  "kind": zod.string(),
  "label": zod.string(),
  "source": zod.string(),
  "xeroAccountId": zod.string().nullish(),
  "xeroAccountName": zod.string().nullish(),
  "sortOrder": zod.number().optional(),
  "createdAt": zod.string().optional()
})).optional()
})


/**
 * @summary Replace all line items for a report
 */
export const ReplaceCustomReportLineItemsParams = zod.object({
  "id": zod.coerce.string()
})

export const ReplaceCustomReportLineItemsBody = zod.object({
  "items": zod.array(zod.object({
  "kind": zod.string(),
  "label": zod.string(),
  "source": zod.string(),
  "xeroAccountId": zod.string().nullish(),
  "xeroAccountName": zod.string().nullish(),
  "sortOrder": zod.number().optional()
})).optional()
})

export const ReplaceCustomReportLineItemsResponse = zod.object({
  "items": zod.array(zod.object({
  "id": zod.string(),
  "reportId": zod.string(),
  "kind": zod.string(),
  "label": zod.string(),
  "source": zod.string(),
  "xeroAccountId": zod.string().nullish(),
  "xeroAccountName": zod.string().nullish(),
  "sortOrder": zod.number().optional(),
  "createdAt": zod.string().optional()
})).optional()
})


/**
 * @summary Get monthly aggregated data for a report
 */
export const GetCustomReportDataParams = zod.object({
  "id": zod.coerce.string()
})

export const GetCustomReportDataResponse = zod.object({
  "months": zod.array(zod.object({
  "month": zod.string(),
  "income": zod.number(),
  "expenses": zod.number(),
  "net": zod.number()
})),
  "totals": zod.object({
  "income": zod.number().optional(),
  "expenses": zod.number().optional(),
  "net": zod.number().optional()
}),
  "growth": zod.object({
  "momPct": zod.number().nullish(),
  "popPct": zod.number().nullish()
}),
  "period": zod.object({
  "fromDate": zod.string().optional(),
  "toDate": zod.string().optional(),
  "months": zod.number().optional()
})
})