# lib/api-zod/src/generated/types/createCustomReportBody.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCustomReportBody {
  cafeId: string;
  name: string;
  description?: string | null;
  dateRangeMonths?: number;
}