# lib/api-zod/src/generated/types/customReport.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CustomReport {
  id: string;
  cafeId: string;
  ownerId?: string;
  name: string;
  description?: string | null;
  dateRangeMonths: number;
  includeInIm: boolean;
  incomeCount?: number;
  expenseCount?: number;
  createdAt?: string;
  updatedAt?: string;
}