# lib/api-zod/src/generated/types/customReportData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CustomReportDataGrowth } from './customReportDataGrowth';
import type { CustomReportDataPeriod } from './customReportDataPeriod';
import type { CustomReportDataTotals } from './customReportDataTotals';
import type { CustomReportMonthRow } from './customReportMonthRow';

export interface CustomReportData {
  months: CustomReportMonthRow[];
  totals: CustomReportDataTotals;
  growth: CustomReportDataGrowth;
  period: CustomReportDataPeriod;
}